+++ powerbox/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "0.2.3"
+__version__ = "0.3.0"
 
 from powerbox import PowerBox, get_power, LogNormalPowerBox